feat: 게시글 작성, 수정, 삭제 시 사용자 검증 로직 추가
- PostController 및 PostService에 인증된 사용자를 활용한 권한 검증 로직 추가
- 인증된 사용자 정보를 위해 Authentication 의존성 적용
- UserNotAllowedException 클래스 생성으로 비허용 사용자 예외 처리
- PostService에서 직접 처리하던 PostEntity 생성 로직을 PostEntity.of() 메서드로 이동
- Post 모델에 User 필드 매핑하여 데이터 응답 확장

+++ thread-board/src/main/java/com/backendstarter/threadboard/service/PostService.java
@@ -1,6 +1,8 @@
 package com.backendstarter.threadboard.service;
 
 import com.backendstarter.threadboard.exception.post.PostNotFoundException;
+import com.backendstarter.threadboard.exception.user.UserNotAllowedException;
+import com.backendstarter.threadboard.model.entity.UserEntity;
 import com.backendstarter.threadboard.model.post.Post;
 import com.backendstarter.threadboard.model.post.PostPatchRequestBody;
 import com.backendstarter.threadboard.model.post.PostPostRequestBody;
@@ -38,30 +40,40 @@ public class PostService {
     }
 
 
-    public Post createPost(PostPostRequestBody postPostRequestBody) {
-        var postEntity = new PostEntity();
-        postEntity.setBody(postPostRequestBody.body());
-        var savedPostEntity = postEntityRepository.save(postEntity);
-
-        return Post.from(savedPostEntity);
+    public Post createPost(PostPostRequestBody postPostRequestBody, UserEntity currentUser) {
+        var postEntity = postEntityRepository.save(
+            PostEntity.of(postPostRequestBody.body(), currentUser)
+        );
+        return Post.from(postEntity);
     }
 
-    public Post updatePost(Long postId, PostPatchRequestBody postPatchRequestBody) {
+    public Post updatePost(Long postId, PostPatchRequestBody postPatchRequestBody, UserEntity currentUser) {
         var postEntity = postEntityRepository
             .findById(postId)
             .orElseThrow(
                 () -> new PostNotFoundException());
+
+        //작성자와 사용자 일치 확인
+        if(!postEntity.getUser().equals(currentUser)) {
+            throw new UserNotAllowedException();
+        }
+
         postEntity.setBody(postPatchRequestBody.body());
         var savedPostEntity = postEntityRepository.save(postEntity);
 
         return Post.from(savedPostEntity);
     }
 
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId, UserEntity currentUser) {
         var postEntity = postEntityRepository
             .findById(postId)
             .orElseThrow(
                 () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Post not found"));
+
+        //작성자와 사용자 일치 확인
+        if(!postEntity.getUser().equals(currentUser)) {
+            throw new UserNotAllowedException();
+        }
 
         postEntityRepository.delete(postEntity);
     }
